git-svn-id: svn://localhost/ardour2/branches/2.1-staging@1698 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ libs/pbd/stacktrace.cc
@@ -23,7 +23,7 @@ PBD::stacktrace (std::ostream& out, int levels)
 
 		printf ("Obtained %zd stack frames.\n", size);
 		
-		for (i = 0; i < size && (levels == 0 || i < levels); i++) {
+		for (i = 0; i < size && (levels == 0 || i < size_t(levels)); i++) {
 			out << strings[i] << std::endl;
 		}
